♻️	refactor: 응답 정규화 Refactoring
