+++ src/main/java/com/example/umc7th/global/apipayload/CustomResponse.java
@@ -2,6 +2,7 @@ package com.example.umc7th.global.apipayload;
 
 
 import com.example.umc7th.global.apipayload.code.BaseSuccessCode;
+import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import lombok.AccessLevel;
@@ -27,14 +28,23 @@ public class CustomResponse<T> {
     private String message;
 
     @JsonProperty("result")
+    @JsonInclude(JsonInclude.Include.NON_NULL)   //
     private T result;
 
 
     public static <T> CustomResponse<T> onSuccess(BaseSuccessCode code, T result) {
         return new CustomResponse(true, HttpStatus.OK, code.getCode(), HttpStatus.OK.getReasonPhrase(), result);
     }
+    // result 없는 성공
+    public static <T> CustomResponse<T> onSuccess(BaseSuccessCode code) {
+        return new CustomResponse(true, HttpStatus.OK, code.getCode(), HttpStatus.OK.getReasonPhrase(), null);
+    }
 
     public static <T> CustomResponse<T> onFailure(HttpStatus status, String code, String message, T result){
         return new CustomResponse(false, status, code, message, result);
+    }
+    // result 없는 실패
+    public static <T> CustomResponse<T> onFailure(HttpStatus status, String code, String message){
+        return new CustomResponse(false, status, code, message, null);
     }
 }
