+++ src/main/java/com/example/umc7th/global/apipayload/code/BaseErrorCode.java
@@ -1,5 +1,6 @@
 package com.example.umc7th.global.apipayload.code;
 
+import com.example.umc7th.global.apipayload.CustomResponse;
 import org.springframework.http.HttpStatus;
 
 public interface BaseErrorCode {
@@ -7,5 +8,9 @@ public interface BaseErrorCode {
     HttpStatus getStatus();
     String getCode();
     String getMessage();
+
+    public default CustomResponse<Void> errorResponse(){
+        return CustomResponse.onFailure(getStatus(), getCode(), getMessage());
+    }
 
 }
